pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["melloru <pythonubuntu12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
